feat: migrate DACBotCheckboxComponent.spec
